--- ram.cpp
@@ -1,57 +1,0 @@
-#include <vector>
-namespace RAM
-  {
-    // how the fuck does the memory even work !!!!??!?!?!?!?
-    std::vector<u8> rom(0x8000); // I have no idea what the size should be
-    std::vector<u8> vRam(0xA000 - 0x8000);  
-    std::vector<u8> iRam(0xC000 - 0xA000);
-    std::vector<u8> mbc (0xE000 - 0xC000); // This size should be big enough to cover all RAM banks
-    std::vector<u8> sRam(0xFE00 - 0xE000);
-    std::vector<u8> OAM (0xFEA0 - 0xFE00);  // this one needs to be made smaller
-    std::vector<u8> i2Ram(0xFFFF - 0xFF80); // more internal ram
-    std::vector<u8> bootRom =  {0x31, 0xfe, 0xff, 0xaf, 0x21, 0xff, 0x9f, 0x32, 0xcb, 0x7c, 0x20, 0xfb, 0x21, 0x26, 0xff, 0x0e, 0x11, 0x3e, 0x80, 0x32, 0xe2, 0x0c, 0x3e, 0xf3, 0xe2, 0x32, 0x3e, 0x77, 0x77, 0x3e, 0xfc, 0xe0, 0x47, 0x11, 0x04, 0x01, 0x21, 0x10, 0x80, 0x1a, 0xcd, 0x95, 0x00, 0xcd, 0x96, 0x00, 0x13, 0x7b, 0xfe, 0x34, 0x20, 0xf3, 0x11, 0xd8, 0x00, 0x06, 0x08, 0x1a, 0x13, 0x22, 0x23, 0x05, 0x20, 0xf9, 0x3e, 0x19, 0xea, 0x10, 0x99, 0x21, 0x2f, 0x99, 0x0e, 0x0c, 0x3d, 0x28, 0x08, 0x32, 0x0d, 0x20, 0xf9, 0x2e, 0x0f, 0x18, 0xf3, 0x67, 0x3e, 0x64, 0x57, 0xe0, 0x42, 0x3e, 0x91, 0xe0, 0x40, 0x04, 0x1e, 0x02, 0x0e, 0x0c, 0xf0, 0x44, 0xfe, 0x90, 0x20, 0xfa, 0x0d, 0x20, 0xf7, 0x1d, 0x20, 0xf2, 0x0e, 0x13, 0x24, 0x7c, 0x1e, 0x83, 0xfe, 0x62, 0x28, 0x06, 0x1e, 0xc1, 0xfe, 0x64, 0x20, 0x06, 0x7b, 0xe2, 0x0c, 0x3e, 0x87, 0xe2, 0xf0, 0x42, 0x90, 0xe0, 0x42, 0x15, 0x20, 0xd2, 0x05, 0x20, 0x4f, 0x16, 0x20, 0x18, 0xcb, 0x4f, 0x06, 0x04, 0xc5, 0xcb, 0x11, 0x17, 0xc1, 0xcb, 0x11, 0x17, 0x05, 0x20, 0xf5, 0x22, 0x23, 0x22, 0x23, 0xc9, 0xce, 0xed, 0x66, 0x66, 0xcc, 0x0d, 0x00, 0x0b, 0x03, 0x73, 0x00, 0x83, 0x00, 0x0c, 0x00, 0x0d, 0x00, 0x08, 0x11, 0x1f, 0x88, 0x89, 0x00, 0x0e, 0xdc, 0xcc, 0x6e, 0xe6, 0xdd, 0xdd, 0xd9, 0x99, 0xbb, 0xbb, 0x67, 0x63, 0x6e, 0x0e, 0xec, 0xcc, 0xdd, 0xdc, 0x99, 0x9f, 0xbb, 0xb9, 0x33, 0x3e, 0x3c, 0x42, 0xb9, 0xa5, 0xb9, 0xa5, 0x42, 0x3c, 0x21, 0x04, 0x01, 0x11, 0xa8, 0x00, 0x1a, 0x13, 0xbe, 0x20, 0xfe, 0x23, 0x7d, 0xfe, 0x34, 0x20, 0xf5, 0x06, 0x19, 0x78, 0x86, 0x23, 0x05, 0x20, 0xfb, 0x86, 0x20, 0xfe, 0x3e, 0x01, 0xe0, 0x50};
-    std::vector<u8> moreRam(0xFF80 - 0xFEA0); //TODO: needs a better name
-    // We'll handle memory properly later
-    u8 readAt(u16 addr)
-    {
-          // we'll be assuming 32kB cartridge for now and add swapping in later
-        if (addr < 0x8000){
-            if (addr < 0x0100){
-                if (moreRam[0xFF50] == 0) // if the bootrom is enabled (which it is by default)
-                {
-                    return bootRom.at(addr);
-                }
-            }
-            return rom.at(addr);
-        }
-        else if(addr < 0xA000)                    { return vRam.at(addr - 0x8000); }
-        else if(addr < 0xC000)                    { return sRam.at(addr - 0xA000); }
-        else if(addr < 0xE000)                    { return iRam.at(addr - 0xC000); }
-        else if(addr < 0xFE00)                    { return iRam.at(addr - 0xE000); }
-        else if(addr < 0xFEA0)                    { return OAM.at(addr - 0xFE00); }
-        else if(addr < 0xFF80)                    { return moreRam.at(addr - 0xFEA0);}
-        else if(0xFF80 <= addr && addr < 0xFFFF)  { return i2Ram.at(addr - 0xFF80); }
-        return -1;
-    }
-
-    void write(u8 val, u16 addr)
-    {
-        if (addr < 0x8000)                        { rom.at(addr) = val; }
-        else if(addr < 0xA000)                    { vRam.at(addr - 0x8000) = val; }
-        else if(addr < 0xC000)                    { sRam.at(addr - 0xA000) = val; }
-        else if(addr < 0xE000)                    { iRam.at(addr - 0xC000) = val; }
-        else if(addr < 0xFE00)                    { iRam.at(addr - 0xE000) = val; }
-        else if(addr < 0xFEA0)                    { OAM.at(addr - 0xFE00) = val; }
-        else if(addr < 0xFF80)                    { moreRam.at(addr - 0xFEA0) = val; }
-        else if(0xFF80 <= addr && addr < 0xFFFF)  { i2Ram.at(addr - 0xFF80) = val; }
-    }
-
-    //could just make this a function call in CPU
-    u8 read()
-    { 
-        u8 byte = readAt(CPU::PC);
-        ++CPU::PC;
-        return byte;
-    }
-
-  }
